# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Citas",
        "description": "Documentación de la API REST para la gestión de citas/reservas en un sistema de peluquería",
        "contact": {
            "email": "jo46646@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Listar todos los usuarios",
                "description": "Obtiene una lista de todos los usuarios activos",
                "operationId": "a5c2f85f763142435e660e6d2fa26c42",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error inesperado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error inesperado. Por favor, inténtalo nuevamente más tarde."
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Crear un nuevo usuario",
                "description": "Registra un nuevo usuario en el sistema",
                "operationId": "99822c6d026ea6da87cb4926e9ee9705",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario. Debe contener al menos una letra minúscula, una mayúscula y un número.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    },
                                    "role_id": {
                                        "description": "ID del rol del usuario (opcional)",
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado con éxito"
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error inesperado. Por favor, inténtalo nuevamente más tarde."
                                        },
                                        "error": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo nombre es obligatorio."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo correo electrónico es obligatorio."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo contraseña debe contener al menos una letra mayúscula."
                                                    }
                                                },
                                                "role_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El rol seleccionado no es válido."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}